resolves first conflicts integrating hmartinezs code

+++ common-libs/src/sw_server.c
@@ -7,8 +7,8 @@
 void sw_server_input_init(list_t* sw_list, list_t* write_list, unsigned int write_size, 
 			  float match, float mismatch, float gap_open, float gap_extend, 
 			  float min_score, unsigned int flank_length, genome_t* genome, 
-			  size_t max_intron_size, size_t min_intron_size, unsigned int seed_max_distance, 
-			  sw_server_input_t* input) {
+			  size_t max_intron_size, size_t min_intron_size, 
+			  size_t seed_max_distance, sw_server_input_t* input) {
   
   input->sw_list_p = sw_list;
   input->write_list_p = write_list;
